Some C++ sorces fixes
JulDaytoDate
    1) Change "ut" from float to double. Is there a reason why "ut"
     should be float, not double? If it must be float, than change other
     numbers to float also.

MidTime
    1) Change some numbers to float type from double

TimeDifference
    1) Change some numbers to float type from double

Unique
    1) Change "pVal" from int to T type. It seems like it should be T.
    2) Change conditions order. If "i == 0" stay first and is True, than
     further conditions will not be checked,so "pVal" will not be used
     before the first initialization in the body for loop

+++ DateTimeTools/__data/libdatetime/JulDaytoDate.cc
@@ -12,10 +12,10 @@
  *
  * OUTPUTS :
  * 			int		*Date		Date array
- * 			float	*ut			Time array in decimal hours 			
+ * 			double	*ut			Time array in decimal hours 			
  * 
  * ********************************************************************/
-void JulDaytoDate(int n, double *JD, int *Date, float *ut) {
+void JulDaytoDate(int n, double *JD, int *Date, double *ut) {
 	
 	/* split the date */
 	int *yr = new int[n];
